Refactor API routes into separate modules
