+++ adhd-focus-hub/backend/api/__init__.py
@@ -1,3 +1,3 @@
 """API package for ADHD Focus Hub."""
 
-__all__ = ["main", "models"]
+__all__ = ["main", "models", "routes"]
